refactor(events): extract RSVP question serializer from helpers
Keep _event_helpers under the 500-line limit after adding rsvp_questions
serialization.
